+++ test/fonttest.cc
@@ -2,6 +2,7 @@
 #include <tuple>
 #include <functional>
 
+#include "../util.h"
 #include "../font.h"
 
 std::vector<std::string> font_names =
