+++ util.cc
@@ -1,0 +1,143 @@
+/* util.cc
+ *   by Trinity Quirk <tquirk@ymb.net>
+ *   last updated 29 Nov 2019, 15:29:33 tquirk
+ *
+ * CuddlyGL OpenGL widget toolkit
+ * Copyright (C) 2019  Trinity Annabelle Quirk
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ *
+ *
+ * This file contains some string conversion functions.
+ *
+ * Things to do
+ *
+ */
+
+#include "util.h"
+
+/* We need to be able to convert from UTF-8 representation to
+ * actual Unicode code points and back.  All network traffic
+ * should be in UTF-8, and we'll of course need to display things
+ * in whatever native font the user needs.
+ *
+ * Ref: http://www.cprogramming.com/tutorial/unicode.html
+ * Ref: https://www.cl.cam.ac.uk/~mgk25/unicode.html
+ */
+std::u32string ui::utf8tou32str(const std::string& str)
+{
+    std::string::const_iterator i = str.begin();
+    std::u32string newstr;
+    uint32_t ch;
+
+    while (i != str.end())
+    {
+        if ((*i & 0xfe) == 0xfc)
+        {
+            ch = (*i & 0x01) << 30;
+            ch |= (*(++i) & 0x3f) << 24;
+            ch |= (*(++i) & 0x3f) << 18;
+            ch |= (*(++i) & 0x3f) << 12;
+            ch |= (*(++i) & 0x3f) << 6;
+            ch |= (*(++i) & 0x3f);
+        }
+        else if ((*i & 0xfc) == 0xf8)
+        {
+            ch = (*i & 0x03) << 24;
+            ch |= (*(++i) & 0x3f) << 18;
+            ch |= (*(++i) & 0x3f) << 12;
+            ch |= (*(++i) & 0x3f) << 6;
+            ch |= (*(++i) & 0x3f);
+        }
+        else if ((*i & 0xf8) == 0xf0)
+        {
+            ch = (*i & 0x07) << 18;
+            ch |= (*(++i) & 0x3f) << 12;
+            ch |= (*(++i) & 0x3f) << 6;
+            ch |= (*(++i) & 0x3f);
+        }
+        else if ((*i & 0xf0) == 0xe0)
+        {
+            ch = (*i & 0x0f) << 12;
+            ch |= (*(++i) & 0x3f) << 6;
+            ch |= (*(++i) & 0x3f);
+        }
+        else if ((*i & 0xe0) == 0xc0)
+        {
+            ch = (*i & 0x1f) << 6;
+            ch |= (*(++i) & 0x3f);
+        }
+        else if ((*i & 0x80) == 0x00)
+            ch = *i;
+        else
+            ch = 0xfffd;
+
+        newstr.push_back(ch);
+        ++i;
+    }
+    return newstr;
+}
+
+std::string ui::u32strtoutf8(const std::u32string& str)
+{
+    std::u32string::const_iterator i = str.begin();
+    std::string newstr;
+
+    while (i != str.end())
+    {
+        if (*i & 0x7c000000)
+        {
+            newstr.push_back(0xfc | ((*i & 0x40000000) >> 30));
+            newstr.push_back(0x80 | ((*i & 0x3f000000) >> 24));
+            newstr.push_back(0x80 | ((*i & 0xfc0000) >> 18));
+            newstr.push_back(0x80 | ((*i & 0x3f000) >> 12));
+            newstr.push_back(0x80 | ((*i & 0xfc0) >> 6));
+            newstr.push_back(0x80 | (*i & 0x3f));
+        }
+        else if (*i & 0x3e00000)
+        {
+            newstr.push_back(0xf8 | ((*i & 0x3000000) >> 24));
+            newstr.push_back(0x80 | ((*i & 0xfc0000) >> 18));
+            newstr.push_back(0x80 | ((*i & 0x3f000) >> 12));
+            newstr.push_back(0x80 | ((*i & 0xfc0) >> 6));
+            newstr.push_back(0x80 | (*i & 0x3f));
+        }
+        else if (*i & 0x1f0000)
+        {
+            newstr.push_back(0xf0 | ((*i & 0x1c0000) >> 18));
+            newstr.push_back(0x80 | ((*i & 0x3f000) >> 12));
+            newstr.push_back(0x80 | ((*i & 0xfc0) >> 6));
+            newstr.push_back(0x80 | (*i & 0x3f));
+        }
+        else if (*i & 0xf800)
+        {
+            newstr.push_back(0xe0 | ((*i & 0x1f000) >> 12));
+            newstr.push_back(0x80 | ((*i & 0xfc0) >> 6));
+            newstr.push_back(0x80 | (*i & 0x3f));
+        }
+        else if (*i & 0x780)
+        {
+            newstr.push_back(0xc0 | ((*i & 0x7c0) >> 6));
+            newstr.push_back(0x80 | (*i & 0x3f));
+        }
+        else
+        {
+            newstr.push_back(*i & 0x7f);
+        }
+        ++i;
+    }
+
+    return newstr;
+}
